Done some renaming in files
